+++ CHANGELOG.md
@@ -17,3 +17,4 @@
     01. :gift: Added class `dev.orne.beans.converters.YearMonthConverter`
     01. :gift: Added class `dev.orne.beans.converters.ZonedDateTimeConverter`
     01. :gift: Added class `dev.orne.beans.converters.ZoneOffsetConverter`
+    01. :gift: Added class `dev.orne.beans.converters.DurationConverter`